+++ lib/modules/setpieces.dart
@@ -199,7 +199,7 @@ class Setpieces extends ChangeNotifier {
                     final localization = Localization();
                     return localization.translate('ui.buttons.leave');
                   }(),
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -236,7 +236,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -265,7 +265,7 @@ class Setpieces extends ChangeNotifier {
                     return localization
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -298,7 +298,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -332,7 +332,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -359,7 +359,7 @@ class Setpieces extends ChangeNotifier {
                     return localization
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -387,7 +387,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -424,7 +424,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -452,7 +452,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -480,7 +480,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -507,7 +507,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -536,7 +536,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -561,7 +561,7 @@ class Setpieces extends ChangeNotifier {
                         .translate('setpieces.cave_scenes.leave_cave');
                   }(),
                   'cooldown': 1,
-                  'nextScene': 'end'
+                  'nextScene': 'leave_end'
                 }
               }
             },
@@ -579,6 +579,14 @@ class Setpieces extends ChangeNotifier {
                   'nextScene': 'finish'
                 }
               }
+            },
+            'leave_end': {
+              'text': () {
+                final localization = Localization();
+                return [localization.translate('setpieces.cave.leave_text')];
+              }(),
+              'onLoad': 'endEvent',
+              'buttons': {}
             }
           },
           'audio': 'landmark_cave'